+++ devdesk/src/App.js
@@ -1,5 +1,5 @@
-import React, { useState } from 'react';
-import {Route, Link} from "react-router-dom";
+import React from "react";
+import { Route } from "react-router-dom";
 import './App.css';
 
 import Header from "./components/Header/Header";
@@ -8,28 +8,18 @@ import Login from "./components/UserOnboarding/Login";
 import StudentDashboard from "./components/Student/StudentDashboard";
 
 function App() {
-  const initialSignupValues = {
-    username: "",
-    password: "",
-    userType: "",
-  };
-
-  const [signupValues, setSignupValues] = useState(initialSignupValues);
-
   return (
     <div className="App">
       <Header />
       <Route exact path="/" render={() => 
           <>
-            <div>Homepage</div><br/>
-            <Link to="/signup/">Signup</Link><br/>
-            <Link to="/login/">Login</Link>
+            <div>Homepage</div>
           </>
         }
       />
-      <Route path="/signup/" render={(props) => <Signup {...props} signupValues={signupValues} />} />
+      <Route path="/signup/" render={(props) => <Signup {...props} />} />
       <Route path="/login/" render={(props) => <Login {...props} />} />
-      <Route path="/student-dashboard" render={(props) => <StudentDashboard {...props}/>} />
+      <Route path="/student-dashboard/" render={(props) => <StudentDashboard {...props}/>} />
     </div>
   );
 }
