+++ devdesk/src/components/UserOnboarding/Signup.js
@@ -1,8 +1,7 @@
-import React from "react";
+import React, { useState } from "react";
 import { Formik, Form, Field, ErrorMessage } from "formik";
 import * as yup from "yup";
 import styled from "styled-components";
-import { Link, Redirect } from "react-router-dom";
 
 const StyledFormikDiv = styled.div`
     background: darkslategrey;
@@ -55,7 +54,13 @@ const StyledFormikDiv = styled.div`
 `;
 
 export default function Signup(props) {
-    const {signupValues} = props;
+    const initialSignupValues = {
+        username: "",
+        password: "",
+        userType: "",
+      };
+    
+    const [signupValues, setSignupValues] = useState(initialSignupValues);
 
     const onSubmit = (formValues, actions) => {
         const userToPost = {
@@ -85,8 +90,6 @@ export default function Signup(props) {
     });
 
     return (
-        <>
-            <Link to="/">Homepage</Link>
             <StyledFormikDiv className="signup-form">
                 <Formik 
                     initialValues={signupValues}
@@ -125,6 +128,5 @@ export default function Signup(props) {
                     }}
                 />
             </StyledFormikDiv>
-        </>
     )
 }
